weight_report page and calories_report page are done with no actual data, also renamed some files to match flutter names criteria.
