+++ lib/core/widgets/custom_button.dart
@@ -1,7 +1,5 @@
 import 'package:egycal/core/utilis/size_config.dart';
-import 'package:egycal/features/welcome/presentation/widgets/Welcome.dart';
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 
 class NextCustomButton extends StatelessWidget {
     final PageController pageController ;
@@ -17,14 +15,14 @@ class NextCustomButton extends StatelessWidget {
         right: 32,
         child: ElevatedButton(
           onPressed: ()=>onPressedfun(),
-          child: Icon(
-            Icons.navigate_next,
-            color: Colors.white,
-          ),
           style: ElevatedButton.styleFrom(
             shape: CircleBorder(),
             padding: EdgeInsets.all(20),
             backgroundColor: Color(0xFF102A43),
+          ),
+          child: Icon(
+            Icons.navigate_next,
+            color: Colors.white,
           ),
         ),
       ),
