--- lib/features/CreateAvatar/presentation/widgets/circlea_vatar.dart
@@ -1,22 +1,0 @@
-import 'package:egycal/features/Home/presentation/widgets/home.dart';
-import 'package:flutter/material.dart';
-
-class avatarimg extends StatelessWidget {
-  final String? image;
-  const avatarimg({super.key , this.image});
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: (){
-        Navigator.push(context,
-        MaterialPageRoute(builder: (context) => Home())
-        );
-      },
-      child: CircleAvatar(
-        radius: 50,
-        backgroundImage:AssetImage(image!),
-      ),
-    );
-  }
-}
