+++ lib/features/height/presentation/widgets/centimeters.dart
@@ -1,22 +1,20 @@
 import 'package:flutter/material.dart';
 
-class centimeter extends StatelessWidget {
-    int cm=10;
+class Centimeter extends StatelessWidget {
+  int cm=10;
 
-  centimeter({super.key, required this.cm});
+  Centimeter({super.key, required this.cm});
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      child: Center(
-        child: Text(
-          cm.toString(),
-          style: TextStyle(
-            fontFamily: 'Inter',
-            fontWeight: FontWeight.normal,
-            fontSize: 16,
-            color: Color(0xFF0D1220),
-          ),
+    return Center(
+      child: Text(
+        cm.toString(),
+        style: TextStyle(
+          fontFamily: 'Inter',
+          fontWeight: FontWeight.normal,
+          fontSize: 16,
+          color: Color(0xFF0D1220),
         ),
       ),
     );
