+++ lib/features/height/presentation/widgets/m_feet.dart
@@ -19,6 +19,5 @@ class MFeet extends StatelessWidget {
         ),
       ),
     );
-    ;
   }
 }
