+++ lib/features/height/presentation/widgets/meters.dart
@@ -7,16 +7,14 @@ class Meters extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      child: Center(
-        child: Text(
-          number.toString(),
-          style: TextStyle(
-            fontFamily: 'Inter',
-            fontWeight: FontWeight.normal,
-            fontSize: 16,
-            color: Color(0xFF0D1220),
-          ),
+    return Center(
+      child: Text(
+        number.toString(),
+        style: TextStyle(
+          fontFamily: 'Inter',
+          fontWeight: FontWeight.normal,
+          fontSize: 16,
+          color: Color(0xFF0D1220),
         ),
       ),
     );
