--- lib/features/reports/reports.dart
@@ -1,3 +1,0 @@
-class reports{
-
-}
