+++ lib/features/weight/presentation/widgets/weight.dart
@@ -1,7 +1,7 @@
 import 'package:egycal/core/utilis/size_config.dart';
-import 'package:egycal/core/widgets/Navigation_buttons.dart';
+import 'package:egycal/core/widgets/navigation_buttons.dart';
 import 'package:egycal/core/widgets/custom_text.dart';
-import 'package:egycal/features/CreateAvatar/presentation/widgets/Avatar.dart';
+import 'package:egycal/features/create_avatar/presentation/widgets/avatar.dart';
 import 'package:egycal/features/weight/presentation/widgets/weight_picker.dart';
 import 'package:flutter/material.dart';
