+++ lib/main.dart
@@ -1,21 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-
-import 'features/splash/presentation/splashview.dart';
-
+import 'features/splash/presentation/splash_view.dart';
 void main() {
-  runApp(EgyCal(
-    
-  ));
+  runApp(EgyCal());
 }
 
 class EgyCal extends StatelessWidget {
-  
+  const EgyCal({super.key});
 
   @override
   Widget build(BuildContext context) {
     return SafeArea(
       child: GetMaterialApp(
+        debugShowCheckedModeBanner: false,
         home: SplashView(),
       ),
     );
