+++ pubspec.yaml
@@ -37,6 +37,8 @@ dependencies:
   get: ^4.6.6
   pinput: ^5.0.1
   smooth_page_indicator: ^1.2.0+3
+  fl_chart: ^0.71.0
+  intl: ^0.20.2
 
 dev_dependencies:
   flutter_lints: ^5.0.0
